moving fa-specifics to fa
