+++ QtAwesome/QtAwesome.py
@@ -4,41 +4,12 @@ PyQtAwesome - use font-awesome in PyQt / PySide applications
 This is a port to Python of the C++ QtAwesome by Rick Blommers
 """
 from __future__ import print_function
-from PyQt4.QtCore import (Qt, QObject, QPoint, QRect, QChar, qRound, 
-                          QByteArray, QIODevice)
-from PyQt4.QtGui import (QFontDatabase, QIcon, QIconEngine, QPainter, QColor,
-                         QPixmap, QFont)
-import os
-from .fa import codes
-
-_default_options = {
-    'color': QColor(50, 50, 50),
-    'color-disabled': QColor(70, 70, 70, 60),
-    'color-active': QColor(10, 10, 10),
-    'color-selected': QColor(10, 10, 10),
-    'scale-factor': 0.9,
-}
-
-_resource = { 'fontname' : None }
-
-def _load():
-    """Load the font file"""
-    if(_resource['fontname'] is None):
-        path = os.path.join(os.path.dirname(os.path.realpath(__file__)),
-                            'fonts', 'fontawesome-4.3.0.ttf')
-        with open(path, 'r') as file:
-            font_data = QByteArray(file.read())
-        
-        id = QFontDatabase.addApplicationFontFromData(font_data)
-        loadedFontFamilies = QFontDatabase.applicationFontFamilies(id)
-        if(loadedFontFamilies):
-            _resource['fontname'] = loadedFontFamilies[0]
-        else:
-            print('Font is empty')
+from PyQt4.QtCore import Qt, QPoint, QRect, qRound
+from PyQt4.QtGui import QIcon, QIconEngine, QPainter, QPixmap
 
 
-class QtAwesomeCharIconPainter: 
-    """The font-awesome icon painter"""
+class CharIconPainter: 
+    """The char icon painter"""
 
     def paint(self, awesome, painter, rect, mode, state, options):
         painter.save()
@@ -66,11 +37,11 @@ class QtAwesomeCharIconPainter:
         painter.restore()
 
 
-class QtAwesomeIconPainterIconEngine(QIconEngine):
+class CharIconEngine(QIconEngine):
     """The painter icon engine"""
 
     def __init__(self, awesome, painter, options):
-        super(QtAwesomeIconPainterIconEngine, self).__init__()
+        super(CharIconEngine, self).__init__()
         self.awesome = awesome
         self.painter = painter
         self.options = options
@@ -83,44 +54,3 @@ class QtAwesomeIconPainterIconEngine(QIconEngine):
         pm.fill(Qt.transparent)
         self.paint(QPainter(pm), QRect(QPoint(0, 0), size), mode, state)
         return pm
-
-
-class QtAwesome(QObject):
-    """The main class for managing icons"""
-    
-    def __init__(self):
-        super(QtAwesome, self).__init__()
-        self.painter = QtAwesomeCharIconPainter()
-        self.painters = {}
-        _load()
-
-    def by_char(self, character, options=None):
-        if options is None:
-            options = {}
-        options = dict(_default_options, **options)
-        options['text'] = QChar(int(character))
-        return self.by_painter(self.painter, options)
-
-    def by_name(self, name, options=None):
-        if name in codes:
-            return self.by_char(codes[name], options)
-        if name in self.painters:
-            painter = self.painters[name]
-            return self.by_painter(painter, options)     
-        else:
-            return QIcon()
-    
-    def by_painter(self, painter, options=None):
-        if options is None:
-            options = {}
-        options = dict(_default_options, **options)
-        engine = QtAwesomeIconPainterIconEngine(self, painter, options)
-        return QIcon(engine)
-    
-    def give(self, name, painter):
-        self.painters[name] = painter
-    
-    def font(self, size):
-        font = QFont(_resource['fontname'])
-        font.setPixelSize(size)
-        return font
