+++ QtAwesome/__init__.py
@@ -1,3 +1,1 @@
-from .QtAwesome import QtAwesome
-from .fa import codes
-from .fa import bunch
+from .fa import QtAwesome, codes, bunch
